refactor: remove class methods
